+++ src/components/project/rlgym-learn/subconfigs/serdes/SerdesConfigEditor.tsx
@@ -1,4 +1,3 @@
-import { useState } from "react";
 import { type PyAnySerdeType, type SerdeTypesModel } from "../../../../../models/rlgym-learn/api";
 import SerdesSelect from "./SerdesSelect";
 
@@ -9,11 +8,58 @@ interface SerdesConfigEditorArgs{
 
 function SerdesConfigEditor({serdesTypesModel, setSerdesTypesModel}:SerdesConfigEditorArgs) {
 
+    const serdeText = (serdeName: string, serdeText: string) => {
+        return <div>
+            <p className="display-6">{serdeName}</p>
+            <p>{serdeText}</p>
+        </div>
+    }
+
     return (
         <>
+            <div className="modal modal-xl" tabIndex={-1} id="serdesHelp">
+                <div className="modal-dialog">
+                    <div className="modal-content bg-dark text-light border-light">
+                    <div className="modal-header">
+                        <h5 className="modal-title">Information about serdes</h5>
+                        <button type="button" className="btn-close btn-close-white" data-bs-dismiss="modal" aria-label="Close"></button>
+                    </div>
+                    <div className="modal-body">
+                        <p className="display-5">What is a serde?</p>
+                        <p>A serde is an object used to serialize and deserialize an object in order to put and read this object from/to shared memory.</p>
+                        
+                        <hr></hr>
+                        <p>Below is the list and an explanation for each serde.</p>
+
+                        <div className="d-grid gap-2" style={{gridTemplateColumns: "1fr 1fr"}}>
+
+                            {serdeText("Float", "Allows you to serialize a float into shared memory.")}
+                            {serdeText("Int", "Allows you to serialize an integer into shared memory.")}
+                            {serdeText("String", "Allows you to serialize a string into shared memory.")}
+                            {serdeText("Boolean", "Allows you to serialize a bool into shared memory.")}
+                            {serdeText("Bytes", "Allows you to \"serialize\" bytes into shared memory.")}
+                            {serdeText("Complex", "Allows you to serialize coimplex numbers into shared memory.")}
+                            {serdeText("Numpy (WIP)", "Allows you to serialize a numpy array into shared memory. You need to specify the data type used in this numpy array in order to serialize it. This serde is incomplete in the GUI for now due to the fact that the arguments are pickled python functions.")}
+                            {serdeText("Tuple", "Allows you to serialize a tuple of elements into shared memory. You need to specify the type of each element, even if they are the same.")}
+                            {serdeText("Set", "Allows you to serialize a sequence of elements of same type into shared memory. You need to specify the type of the items.")}
+                            {serdeText("Typed dictionary", "Allows you to serialize a specific architecture of dictionary. You can for example, specify you want a String serde for a key called'Key1' and a Numpy serde for another key called 'Key2'.")}
+                            {serdeText("Dictionary", "Allows you to serialize a python dictionary into shared memory. The keys must be of the same type, and so on for the values. You need to specify both the key and value types.")}
+                            {serdeText("Option", "Allows you to serialize an optional field into shared memory. You need to specify the type of the value if given.")}
+                            {serdeText("Union", "Not implemented in this GUI, the only reason is because it requires a python function to be used, like the Numpy serde. Allows you to serialize an object that can be of different types. You need to specify each type.")}
+
+                        </div>
+                    </div>
+                    <div className="modal-footer">
+                        <button type="button" className="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+                    </div>
+                    </div>
+                </div>
+                </div>
             <div className="d-flex mt-2">
                 <p className="display-6">Serdes config</p>
             </div>
+
+            <button className="btn btn-outline-light bi bi-info-circle m-auto" data-bs-toggle="modal" data-bs-target="#serdesHelp"> Help</button>
 
             <div className="d-grid gap-3" style={{gridTemplateColumns: "1fr 1fr"}}>
                 <div>