+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A button to change the data folder, making it possible to change folder during the app runtime
 - There is now a home page!
 - Some description for the project list
+- A modal that explains what a serde is and what it does.
 
 ## [0.1.2] - 2026-06-02